chore: enable hmr for renderer
