+++ package.json
@@ -16,7 +16,7 @@
     "start": "npm run build && run-electron build/electron/main.js --remote-debugging-port=9222 --inspect=9223",
     "start:clean": "npm run clean && npm run start",
     "precommit": "lint-staged",
-    "watch": "npm run build && concurrently \"run-electron build/electron/main.js --remote-debugging-port=9222 --inspect=9223\" \"tsc --project . --watch\" \"webpack -w\"",
+    "watch": "npm run build && concurrently \"run-electron build/electron/main.js --hot --remote-debugging-port=9222 --inspect=9223\" \"tsc --project . --watch\" \"webpack -w\" \"webpack-dev-server --config webpack.renderer.js\"",
     "watch:clean": "npm run clean && npm run watch",
     "watch:tests": "jest --watch",
     "test": "jest"
@@ -148,7 +148,8 @@
     "tslint": "5.11.0",
     "typedoc": "0.11.1",
     "typescript-styled-plugin": "0.6.3",
-    "webpack-cli": "3.0.8"
+    "webpack-cli": "3.0.8",
+    "webpack-dev-server": "3.1.5"
   },
   "dependencies": {
     "@brainly/html-sketchapp": "3.3.1",
@@ -215,7 +216,9 @@
     "username": "3.0.0",
     "uuid": "3.2.1",
     "webpack": "4.6.0",
-    "ws": "5.1.1"
+    "webpack-dev-middleware": "3.1.3",
+    "ws": "5.1.1",
+    "yargs-parser": "10.1.0"
   },
   "homepage": "https://meetalva.github.io/"
 }
