+++ src/electron/main.ts
@@ -3,6 +3,7 @@ import * as Mobx from 'mobx';
 import * as Path from 'path';
 import { AppContext, startApp as start } from './start-app';
 
+const yargsParser = require('yargs-parser');
 const clearModule = require('clear-module');
 const importFresh = require('import-fresh');
 
@@ -11,10 +12,14 @@ const CONTEXT: AppContext = Mobx.observable({
 	port: undefined,
 	project: undefined,
 	sender: undefined,
-	win: undefined
+	win: undefined,
+	hot: undefined
 });
 
 async function main(): Promise<void> {
+	const args = yargsParser(process.argv.slice(2));
+	CONTEXT.hot = args.hot || false;
+
 	const StartApp = importFresh('./start-app');
 	const startApp = StartApp.startApp as typeof start;
 	const app = await startApp(CONTEXT);
