+++ webpack.config.js
@@ -1,17 +1,12 @@
 const Path = require('path');
-const MonacoWebpackPlugin = require('monaco-editor-webpack-plugin');
-const webpack = require('webpack');
 
 const MOBX_PATH = require.resolve('mobx');
-const HTML_SKETCHAPP_PATH = require.resolve('@brainly/html-sketchapp');
-
 const PREVIEW_PATH = require.resolve('./src/preview/preview.ts');
 const PREVIEW_RENDERER_PATH = require.resolve('./src/preview-renderer/index.ts');
 const EXPORT_TO_SKETCH_DATA_PATH = require.resolve('./src/preview/export-to-sketch-data.ts');
 
-const RENDERER_PATH = require.resolve('./src/renderer/index.tsx');
-
 module.exports = [
+	require('./webpack.renderer'),
 	{
 		mode: 'development',
 		entry: {
@@ -42,41 +37,6 @@ module.exports = [
 			library: '[name]',
 			libraryTarget: 'window',
 			path: Path.join(__dirname, 'build', 'scripts')
-		}
-	},
-	{
-		mode: 'development',
-		entry: {
-			renderer: RENDERER_PATH
-		},
-		module: {
-			rules: [
-				{
-					test: /\.css$/,
-					use: ['style-loader', 'css-loader']
-				},
-				{
-					test: /\.tsx?$/,
-					loader: 'ts-loader',
-					options: {
-						transpileOnly: true
-					}
-				}
-			]
-		},
-		plugins: [
-			new MonacoWebpackPlugin()
-		],
-		resolve: {
-			extensions: ['.ts', '.tsx', '.js']
-		},
-		node: {
-			__dirname: false
-		},
-		output: {
-			filename: '[name].js',
-			path: Path.join(__dirname, 'build', 'scripts'),
-			publicPath: '/scripts/'
 		}
 	},
 	{
